+++ lib/src/blocs/page/pop_event.dart
@@ -1,5 +1,6 @@
 import '../page_stack/bloc.dart';
 import 'event.dart';
+import 'pop_cause.dart';
 
 /// Emit this for [PageStackBloc] to remove and dispose the page.
 ///
@@ -7,9 +8,11 @@ import 'event.dart';
 /// is left in the stack afterwards.
 class PageBlocPopEvent<R> extends PageBlocEvent {
   final R? data;
+  final PopCause cause;
 
   const PageBlocPopEvent({
     required this.data,
+    required this.cause,
   });
 }