+++ lib/src/blocs/page_stack/page_factory_function.dart
@@ -1,0 +1,8 @@
+import '../../pages/abstract.dart';
+import '../page/path.dart';
+
+typedef CPageFactoryFunction<P extends PagePath> = CAbstractPage<P, dynamic>?
+    Function(
+  String factoryKey,
+  Map<String, dynamic> state,
+);