+++ lib/src/pages/bloc_material.dart
@@ -9,7 +9,7 @@ import 'material.dart';
 /// [P] is the base class for all app's page path.
 /// [R] is the result returned when the page pops.
 /// [B] is the block for the state.
-abstract class CBlocMaterialPage<P extends PagePath, R,
+class CBlocMaterialPage<P extends PagePath, R,
     B extends CPageBloc<P, R>> extends CAbstractMaterialPage<P, R> {
   @override
   final B bloc;
@@ -29,6 +29,7 @@ abstract class CBlocMaterialPage<P extends PagePath, R,
   @override
   void dispose() {
     bloc.dispose();
+    super.dispose();
   }
 }